Separate packages so each maps to a single future artifact
Move classes so that no package contains classes destined for two
different artifacts, in preparation for splitting the repo into five
modules. The intended package -> artifact mapping is now:

- ome-zarr-imglib2: pyramid, pyramid.backend, pyramid.exceptions,
  pyramid.metadata
- ome-zarr-n5: pyramid.backend.n5
- ome-zarr-zarr-java: pyramid.backend.zarrjava
- fiji: pyramid.fiji, pyramid.fiji.open, pyramid.fiji.open.exceptions,
  pyramid.fiji.util
- fiji-ui: open, open.options, plugins, plugins.settings, ui, ui.util,
  util
