+++ src/main/java/sc/fiji/ome/zarr/pyramid/fiji/PyramidalDataset.java
@@ -44,7 +44,6 @@ import net.imagej.axis.DefaultLinearAxis;
 import net.imglib2.type.NativeType;
 import net.imglib2.type.numeric.RealType;
 
-import sc.fiji.ome.zarr.pyramid.Pyramidal;
 import sc.fiji.ome.zarr.pyramid.backend.PyramidContents;
 import sc.fiji.ome.zarr.pyramid.metadata.AxisCalibration;
