+++ src/main/java/sc/fiji/ome/zarr/pyramid/fiji/open/exceptions/NotASingleScaleImageException.java
@@ -26,24 +26,18 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.open.exceptions;
+package sc.fiji.ome.zarr.pyramid.fiji.open.exceptions;
 
-public class NonExistingResolutionLevelException extends RuntimeException
+public class NotASingleScaleImageException extends RuntimeException
 {
-	public NonExistingResolutionLevelException( final int resolutionLevel, final int numResolutionLevels )
+
+	public NotASingleScaleImageException( final String path )
 	{
-		super( message( resolutionLevel, numResolutionLevels ) );
+		super( "The dataset at path '" + path + "' is not a valid OME-Zarr single scale image." );
 	}
 
-	public NonExistingResolutionLevelException( final int resolutionLevel, final int numResolutionLevels, final Throwable cause )
+	public NotASingleScaleImageException( final String path, final Throwable cause )
 	{
-		super( message( resolutionLevel, numResolutionLevels ), cause );
-	}
-
-	private static String message( final int resolutionLevel, final int numResolutionLevels )
-	{
-		return "Resolution level " + resolutionLevel + " does not exist. "
-				+ "The pyramid has " + numResolutionLevels + " level"
-				+ ( numResolutionLevels == 1 ? "" : "s" ) + " (0–" + ( numResolutionLevels - 1 ) + ").";
+		super( "The dataset at path '" + path + "' is not a valid OME-Zarr single scale image. Cause: " + cause.getMessage(), cause );
 	}
 }
