+++ src/test/java/sc/fiji/ome/zarr/plugins/OpenInBDVCommandTest.java
@@ -47,7 +47,7 @@ import org.scijava.Context;
 import org.scijava.command.CommandService;
 
 import sc.fiji.ome.zarr.open.ZarrOpenActions;
-import sc.fiji.ome.zarr.pyramid.Pyramidal;
+import sc.fiji.ome.zarr.pyramid.fiji.Pyramidal;
 import sc.fiji.ome.zarr.pyramid.fiji.PyramidalService;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
@@ -65,7 +65,7 @@ class OpenInBDVCommandTest
 
 	/**
 	 * Opening an IJ2 dataset in BDV via the {@link OpenInBDVCommand} should produce a second, distinct
-	 * {@link sc.fiji.ome.zarr.pyramid.Pyramidal} that shares the same {@link sc.fiji.ome.zarr.pyramid.backend.PyramidContents}.
+	 * {@link sc.fiji.ome.zarr.pyramid.fiji.Pyramidal} that shares the same {@link sc.fiji.ome.zarr.pyramid.backend.PyramidContents}.
 	 */
 	@Test
 	void runAfterIj2OpenCreatesSecondBdvDataset() throws URISyntaxException, ExecutionException, InterruptedException
