+++ src/test/java/sc/fiji/ome/zarr/pyramid/fiji/PyramidalDatasetAsImagePlusTest.java
@@ -26,55 +26,69 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid;
+package sc.fiji.ome.zarr.pyramid.fiji;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.net.URISyntaxException;
 import java.nio.file.Path;
 
 import org.junit.jupiter.api.Test;
 import org.scijava.Context;
+import org.scijava.convert.ConvertService;
 
+import ij.ImagePlus;
+import net.imagej.Dataset;
 import sc.fiji.ome.zarr.pyramid.backend.PyramidContents;
 import sc.fiji.ome.zarr.pyramid.backend.n5.N5PyramidBackend;
-import sc.fiji.ome.zarr.pyramid.fiji.PyramidalDataset;
 import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
 /**
- * Tests how {@link PyramidalDataset} names itself: a multi-resolution image
- * gets a " (R)" suffix (at any resolution level), while a single-resolution
- * image does not.
+ * Demonstrates how to open an OME-Zarr dataset and obtain a legacy ImageJ1
+ * {@link ImagePlus} handle from it. The intermediate IJ2 {@link Dataset} is
+ * converted to an {@code ImagePlus} via SciJava's {@link ConvertService};
+ * the converter itself is provided by {@code imagej-legacy} dependency on the
+ * classpath.
  */
-class PyramidalDatasetNamingTest
+class PyramidalDatasetAsImagePlusTest
 {
-	private static final String MULTI_LEVEL_RESOURCE = "sc/fiji/ome/zarr/util/5d_testing/5d_dataset_v5.ome.zarr";
-
-	private static final String SINGLE_LEVEL_RESOURCE =
-			"sc/fiji/ome/zarr/util/single_resolution_testing/single_resolution_dataset_v5.ome.zarr";
-
 	@Test
-	void multiResolutionImageNameEndsWithR() throws URISyntaxException
+	void testOpenAsImagePlusConvertService() throws URISyntaxException
 	{
-		final Path path = ZarrTestUtils.resourcePath( MULTI_LEVEL_RESOURCE );
+		final Path path = ZarrTestUtils.resourcePath( "sc/fiji/ome/zarr/util/5d_testing/5d_dataset_v5.ome.zarr" );
+
 		try (Context context = new Context())
 		{
 			final PyramidContents< ? > contents = new N5PyramidBackend().load( path.toUri() );
-			assertTrue( new PyramidalDataset( context, contents, 0 ).getName().endsWith( " (R)" ) );
-			assertTrue( new PyramidalDataset( context, contents, 1 ).getName().endsWith( " (R)" ) );
+			final PyramidalDataset dataset = new PyramidalDataset( context, contents, 0 );
+			final ImagePlus imagePlus = dataset.asImagePlus();
+
+			assertNotNull( imagePlus );
+			// order of dimensions for imagePlus: width, height, channels, slices, frames
+			assertArrayEquals( new int[] { 64, 64, 1, 16, 4 }, imagePlus.getDimensions() );
+			assertEquals( dataset.getName(), imagePlus.getTitle() );
 		}
 	}
 
 	@Test
-	void singleResolutionImageNameHasNoR() throws URISyntaxException
+	void testOpenAsImagePlusSourceAndConverter() throws URISyntaxException
 	{
-		final Path path = ZarrTestUtils.resourcePath( SINGLE_LEVEL_RESOURCE );
+		final Path path = ZarrTestUtils.resourcePath( "sc/fiji/ome/zarr/util/5d_testing/5d_dataset_v5.ome.zarr" );
+
 		try (Context context = new Context())
 		{
 			final PyramidContents< ? > contents = new N5PyramidBackend().load( path.toUri() );
-			final String name = new PyramidalDataset( context, contents, 0 ).getName();
-			assertFalse( name.contains( "(R)" ), "expected no '(R)' in name but was: " + name );
+			ImagePlus imagePlus = new PyramidalDataset( context, contents, 0 ).asImagePlus();
+
+			assertNotNull( imagePlus );
+			// order of dimensions for imagePlus: width, height, channels, slices, frames
+			assertArrayEquals( new int[] { 64, 64, 1, 16, 4 }, imagePlus.getDimensions() );
+			imagePlus = new PyramidalDataset( context, contents, 1 ).asImagePlus();
+			assertNotNull( imagePlus );
+			// order of dimensions for imagePlus: width, height, channels, slices, frames
+			assertArrayEquals( new int[] { 32, 32, 1, 8, 4 }, imagePlus.getDimensions() );
 		}
 	}
 }
